requirements.txt
runpod
requests
Pillow